requirements.txt
aiofiles==23.2.1
aiohappyeyeballs==2.6.1
aiohttp==3.11.14
aiosignal==1.3.2
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.9.0
attrs==25.3.0
audiocraft==1.3.0
audioread==3.0.1
av==11.0.0
blis==1.2.0
cachetools==5.5.2
catalogue==2.0.10
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
cloudpathlib==0.21.0
cloudpickle==3.1.1
colorama==0.4.6
colorlog==6.9.0
confection==0.1.5
cymem==2.0.11
dataclasses-json==0.6.7
decorator==5.2.1
demucs==4.0.1
distro==1.9.0
dnspython==2.7.0
docopt==0.6.2
dora_search==0.1.12
dotenv==0.9.9
einops==0.8.1
email_validator==2.2.0
encodec==0.1.1
faiss-cpu==1.10.0
fastapi==0.115.12
fastapi-cli==0.0.7
ffmpy==0.5.0
filelock==3.18.0
flashy==0.0.2
frozenlist==1.5.0
fsspec==2025.3.0
google-ai-generativelanguage==0.6.15
google-api-core==2.24.2
google-api-python-client==2.166.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-generativeai==0.8.4
googleapis-common-protos==1.69.2
gradio==5.23.1
gradio_client==1.8.0
greenlet==3.1.1
groovy==0.1.2
groq==0.20.0
grpcio==1.71.0
grpcio-status==1.71.0
h11==0.14.0
httpcore==1.0.7
httplib2==0.22.0
httptools==0.6.4
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.29.3
hydra-colorlog==1.2.0
hydra-core==1.3.2
idna==3.10
imageio==2.37.0
imageio-ffmpeg==0.6.0
Jinja2==3.1.6
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
julius==0.2.7
lameenc==1.8.1
langchain==0.3.21
langchain-community==0.3.20
langchain-core==0.3.48
langchain-groq==0.3.1
langchain-text-splitters==0.3.7
langcodes==3.5.0
langsmith==0.3.18
language_data==1.3.0
lazy_loader==0.4
librosa==0.11.0
lightning-utilities==0.14.2
llvmlite==0.44.0
marisa-trie==1.2.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.26.1
mdurl==0.1.2
motor==3.7.0
moviepy==2.1.2
mpmath==1.3.0
msgpack==1.1.0
multidict==6.2.0
murmurhash==1.0.12
mypy-extensions==1.0.0
networkx==3.4.2
num2words==0.5.14
numba==0.61.0
numpy==1.26.4
omegaconf==2.3.0
openunmix==1.3.0
orjson==3.10.16
packaging==24.2
pandas==2.2.3
pillow==10.4.0
platformdirs==4.3.7
pooch==1.8.2
preshed==3.0.9
proglog==0.1.10
propcache==0.3.0
proto-plus==1.26.1
protobuf==5.29.4
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.10.6
pydantic-settings==2.8.1
pydantic_core==2.27.2
pydub==0.25.1
Pygments==2.19.1
pymongo==4.11.3
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
retrying==1.3.4
rich==13.9.4
rich-toolkit==0.13.2
rsa==4.9
ruff==0.11.2
safehttpx==0.1.6
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.15.2
semantic-version==2.10.0
sentence-transformers==3.4.1
sentencepiece==0.2.0
shellingham==1.5.4
six==1.17.0
smart-open==7.1.0
sniffio==1.3.1
soundfile==0.13.1
soxr==0.5.0.post1
spacy==3.8.4
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==2.0.39
srsly==2.5.1
starlette==0.46.1
submitit==1.5.2
sympy==1.13.1
tenacity==9.0.0
thinc==8.3.4
threadpoolctl==3.6.0
tokenizers==0.21.1
tomlkit==0.13.2
torch==2.6.0
torchaudio==2.6.0
torchdata==0.7.0
torchmetrics==1.7.0
torchtext==0.16.0
torchvision==0.16.0
tqdm==4.67.1
transformers==4.50.1
treetable==0.2.5
typer==0.15.2
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2025.2
uritemplate==4.1.1
urllib3==2.3.0
uvicorn==0.34.0
wasabi==1.1.3
watchfiles==1.0.4
weasel==0.4.1
websockets==15.0.1
wrapt==1.17.2
xformers==0.0.22.post7
yarl==1.18.3
zstandard==0.23.0
